(Minimum) Changes to avoid floating-point exceptions   * If coreneuron is compiled with MPI but run without --mpi     flag then coreneuron crashes. This doesn't allow NEURON     to run existing tests (e.g. using python API)   * Instead of using MPI function, using wrappers from mpispike.cpp   * Updated all reduce wrapper to accept size_t / unsigned long int     than just size_t   * Update cast and summary printf
Further cleanup will be done in separate PR, see #490

+++ coreneuron/utils/nrn_stats.cpp
@@ -30,23 +30,23 @@ const int NUM_STATS = 13;
 enum event_type { enq = 0, spike, ite };
 
 void report_cell_stats(void) {
-    long stat_array[NUM_STATS] = {0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0}, gstat_array[NUM_STATS];
+    size_t stat_array[NUM_STATS] = {0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0}, gstat_array[NUM_STATS];
 
     for (int ith = 0; ith < nrn_nthread; ++ith) {
-        stat_array[0] += (long) nrn_threads[ith].ncell;           // number of cells
-        stat_array[10] += (long) nrn_threads[ith].end;            // number of compartments
-        stat_array[1] += (long) nrn_threads[ith].n_presyn;        // number of presyns
-        stat_array[2] += (long) nrn_threads[ith].n_input_presyn;  // number of input presyns
-        stat_array[3] += (long) nrn_threads[ith].n_netcon;        // number of netcons, synapses
-        stat_array[4] += (long) nrn_threads[ith].n_pntproc;       // number of point processes
+        stat_array[0] += nrn_threads[ith].ncell;           // number of cells
+        stat_array[10] += nrn_threads[ith].end;            // number of compartments
+        stat_array[1] += nrn_threads[ith].n_presyn;        // number of presyns
+        stat_array[2] += nrn_threads[ith].n_input_presyn;  // number of input presyns
+        stat_array[3] += nrn_threads[ith].n_netcon;        // number of netcons, synapses
+        stat_array[4] += nrn_threads[ith].n_pntproc;       // number of point processes
         if (nrn_partrans::transfer_thread_data_) {
             size_t n = nrn_partrans::transfer_thread_data_[ith].tar_indices.size();
-            stat_array[11] += (long) n;  // number of transfer targets
+            stat_array[11] += n;  // number of transfer targets
             n = nrn_partrans::transfer_thread_data_[ith].src_indices.size();
-            stat_array[12] += (long) n;  // number of transfer sources
+            stat_array[12] += n;  // number of transfer sources
         }
     }
-    stat_array[5] = (long) spikevec_gid.size();  // number of spikes
+    stat_array[5] = spikevec_gid.size();  // number of spikes
 
     int spikevec_positive_gid_size = 0;
     for (std::size_t i = 0; i < spikevec_gid.size(); ++i) {
@@ -55,7 +55,7 @@ void report_cell_stats(void) {
         }
     }
 
-    stat_array[6] = (long) spikevec_positive_gid_size;  // number of non-negative gid spikes
+    stat_array[6] = spikevec_positive_gid_size;  // number of non-negative gid spikes
 
 #if NRNMPI
     nrnmpi_long_allreduce_vec(stat_array, gstat_array, NUM_STATS, 1);
@@ -66,16 +66,16 @@ void report_cell_stats(void) {
 
     if (nrnmpi_myid == 0 && !corenrn_param.is_quiet()) {
         printf("\n\n Simulation Statistics\n");
-        printf(" Number of cells: %ld\n", gstat_array[0]);
-        printf(" Number of compartments: %ld\n", gstat_array[10]);
-        printf(" Number of presyns: %ld\n", gstat_array[1]);
-        printf(" Number of input presyns: %ld\n", gstat_array[2]);
-        printf(" Number of synapses: %ld\n", gstat_array[3]);
-        printf(" Number of point processes: %ld\n", gstat_array[4]);
-        printf(" Number of transfer sources: %ld\n", gstat_array[12]);
-        printf(" Number of transfer targets: %ld\n", gstat_array[11]);
-        printf(" Number of spikes: %ld\n", gstat_array[5]);
-        printf(" Number of spikes with non negative gid-s: %ld\n", gstat_array[6]);
+        printf(" Number of cells: %zu\n", gstat_array[0]);
+        printf(" Number of compartments: %zu\n", gstat_array[10]);
+        printf(" Number of presyns: %zu\n", gstat_array[1]);
+        printf(" Number of input presyns: %zu\n", gstat_array[2]);
+        printf(" Number of synapses: %zu\n", gstat_array[3]);
+        printf(" Number of point processes: %zu\n", gstat_array[4]);
+        printf(" Number of transfer sources: %zu\n", gstat_array[12]);
+        printf(" Number of transfer targets: %zu\n", gstat_array[11]);
+        printf(" Number of spikes: %zu\n", gstat_array[5]);
+        printf(" Number of spikes with non negative gid-s: %zu\n", gstat_array[6]);
     }
 }
 }  // namespace coreneuron
